requirements.txt
Django==4.1.13
psycopg2-binary==2.9.9
sqlparse==0.2.4
Pillow==10.2.0
djongo==1.3.6
pytz == 2023.4
pymongo==3.12.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
requests==2.31.0
djangorestframework-simplejwt==5.3.1 